+++ src/js/SVGBuilder.d.ts
@@ -10,7 +10,8 @@
  * @modified 2019-11-07 Added the 'Triangle' style class.
  * @modified 2019-11-13 Added the <?xml ...?> tag.
  * @modified 2020-03-25 Ported this class from vanilla-JS to Typescript.
- * @version  1.0.3
+ * @modified 2020-12-17 Added Circle and CircleSection style classes.
+ * @version  1.0.4
  **/
 import { Vertex } from "./Vertex";
 import { SVGSerializable } from "./interfaces";