+++ src/ts/CircleSector.ts
@@ -6,9 +6,6 @@
 
 
 import { Circle } from "./Circle";
-// import { Vector } from "./Vector";
-// import { VertTuple } from "./VertTuple";
-import { Vertex } from "./Vertex";
 import { SVGPathParams, SVGSerializable, XYCoords } from "./interfaces";
 
 // [ 'A', radiusx, radiusy, rotation=0, largeArcFlag=1|0, sweepFlag=0, endx, endy ]
@@ -19,7 +16,6 @@ export type SVGArcPathParams = [ string, number, number, number, number, number,
  *
  * @requires Line
  * @requires SVGSerializale
- * @requires Vertex
  * @requires XYCoords
  **/
 export class CircleSector implements SVGSerializable {
@@ -80,20 +76,20 @@ export class CircleSector implements SVGSerializable {
     toSVGString( options:{className?:string } ) {
 	options = options || {};
 	var buffer : Array<string> = [];
-	/* buffer.push( '<circle' );
+	buffer.push( '<path ' );
 	if( options.className )
 	    buffer.push( ' class="' + options.className + '"' );
-	buffer.push( ' cx="' + this.center.x + '"' );
-	buffer.push( ' cy="' + this.center.y + '"' );
-	buffer.push( ' r="' + this.radius + '"' );
-	buffer.push( ' />' ); */
+	const data : SVGPathParams = CircleSector.circleSectorUtils.describeSVGArc(
+	    this.circle.center.x, this.circle.center.y,
+	    this.circle.radius,
+	    this.startAngle, this.endAngle );
+	buffer.push( ' d="' + data.join(" ") + '" />' );
 	return buffer.join('');
     };
 
     static circleSectorUtils = {
 	/**
 	 * Helper function to convert polar circle coordinates to cartesian coordinates.
-	 * Found at: https://stackoverflow.com/questions/5736398/how-to-calculate-the-svg-path-for-an-arc-of-a-circle
 	 *
 	 * TODO: generalize for ellipses (two radii).
 	 *
@@ -108,31 +104,45 @@ export class CircleSector implements SVGSerializable {
 
 	/**
 	 * Helper function to convert a circle section as SVG arc params (for the `d` attribute).
+	 * Found at: https://stackoverflow.com/questions/5736398/how-to-calculate-the-svg-path-for-an-arc-of-a-circle
 	 *
 	 * TODO: generalize for ellipses (two radii).
 	 *
+	 * @param {boolean} options.moveToStart - If false (default=true) the initial 'Move' command will not be used.
 	 * @return [ 'A', radiusx, radiusy, rotation=0, largeArcFlag=1|0, sweepFlag=0, endx, endy ]
 	 */
-	describeSVGArc : ( x:number, y:number, radius:number, startAngle:number, endAngle:number ) : SVGPathParams => {
+	describeSVGArc : ( x:number, y:number,
+			   radius:number,
+			   startAngle:number, endAngle:number,
+			   options?:{moveToStart:boolean}
+			 ) : SVGPathParams => {
+	    
+	    if( typeof options === 'undefined' )
+		options = { moveToStart : true };
+	    
 	    const end : XYCoords = CircleSector.circleSectorUtils.polarToCartesian(x, y, radius, endAngle);
 	    const start : XYCoords = CircleSector.circleSectorUtils.polarToCartesian(x, y, radius, startAngle);
 
 	    // Split full circles into two halves.
 	    // Some browsers have problems to render full circles (described by start==end).
 	    if( Math.PI*2-Math.abs(startAngle-endAngle) < 0.001 ) {
-		const firstHalf : SVGPathParams = CircleSector.circleSectorUtils.describeSVGArc( x, y, radius, startAngle, startAngle+(endAngle-startAngle)/2 );
-		// const firstEndPoint : XYCoords = new Vertex( firstHalf[firstHalf.length-2], firstHalf[firstHalf.length-1] );
+		const firstHalf : SVGPathParams = CircleSector.circleSectorUtils.describeSVGArc( x, y, radius, startAngle, startAngle+(endAngle-startAngle)/2, options );
 		const firstEndPoint : XYCoords = { x : firstHalf[firstHalf.length-2] as number,
 						   y : firstHalf[firstHalf.length-1] as number
 						 };
-		const secondHalf : SVGPathParams = CircleSector.circleSectorUtils.describeSVGArc( x, y, radius, startAngle+(endAngle-startAngle)/2, endAngle );
+		const secondHalf : SVGPathParams = CircleSector.circleSectorUtils.describeSVGArc( x, y, radius, startAngle+(endAngle-startAngle)/2, endAngle, options );
 		return firstHalf.concat( secondHalf );
 	    }
 
 	    // Boolean stored as integers (0|1).
 	    const largeArcFlag : number = endAngle - startAngle <= Math.PI ? 0 : 1;
 	    const sweepFlag : number = 1;
-	    return ["A", radius, radius, 0, largeArcFlag, sweepFlag, end.x, end.y ];
+	    const pathData = [];
+	    if( options.moveToStart ) {
+		pathData.push('M', start.x, start.y );
+	    }
+	    pathData.push("A", radius, radius, 0, largeArcFlag, sweepFlag, end.x, end.y );
+	    return pathData;
 	}
     };